requirements.txt
Flask
pandas
scikit-learn
pymongo
python-dotenv
